+++ src/main/java/com/dev/voucherproject/config/console/MenuControllerProvider.java
@@ -1,4 +1,4 @@
-package com.dev.voucherproject.model.menu;
+package com.dev.voucherproject.config.console;
 
 
 import com.dev.voucherproject.controller.console.MenuController;
